move data manager es helpers
